requirements.txt
Flask>=2.0
python-dotenv
aiogram>=3.0
python-dotenv>=0.21
